requirements.txt
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
notion-client
python-dotenv
google-generativeai
beautifulsoup4
html2text
